# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wedding guest-list backend. Collaborative RSVP management with\nrole-based access (owner/editor/viewer), invite links, stats,\nactivity feed, notifications, and CSV export.\n\nAll responses use the envelope `{ \"data\": ..., \"error\": null }`.\nOn error, `data` is null and `error` is `{ \"code\", \"message\", \"details\" }`.\nAuthenticate via `POST /auth/login`, then send `Authorization: Bearer \u003caccess\u003e`.",
        "title": "Zawaj API",
        "contact": {
            "name": "Zawaj"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/weddings/{id}/activity": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated, newest-first activity for a wedding.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Activity feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wedding ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 50, max 200)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ items: [...] }",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "zawaj_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "zawaj_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/zawaj_pkg_response.APIError"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}